# testsuite: Add dircache-focused performance tests to afp_lantest