--- a/.dockerignore
+++ b/.dockerignore
@@ -1,0 +1,5 @@
+# Build artifacts
+build/
+clean/
+*.o
+*.a