--- a/doc/manpages/man1/meson.build
+++ b/doc/manpages/man1/meson.build
@@ -1,5 +1,6 @@
 manfiles = [
     'ad',
+    'afp_lantest',
     'afpldaptest',
     'afppasswd',
     'afpstats',