--- a/doc/po/afp_lantest.1.md.ja.po
+++ b/doc/po/afp_lantest.1.md.ja.po
@@ -1,0 +1,540 @@
+# Japanese translations for Netatalk package
+# Copyright (C) 2025 Contributors to the Netatalk Project
+# This file is distributed under the same license as the Netatalk package.
+# Automatically generated, 2025.
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Netatalk 4.3.0\n"
+"POT-Creation-Date: 2025-08-04 14:23+1000\n"
+"PO-Revision-Date: 2025-08-04 14:22+1000\n"
+"Last-Translator: Automatically generated\n"
+"Language-Team: none\n"
+"Language: ja\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: nplurals=1; plural=0;\n"
+
+#. type: Plain text
+#: manpages/man1/ad.1.md:2 manpages/man1/addump.1.md:2
+#: manpages/man1/aecho.1.md:2 manpages/man1/afp_lantest.1.md:2
+#: manpages/man1/afpldaptest.1.md:2 manpages/man1/afppasswd.1.md:2
+#: manpages/man1/afpstats.1.md:2 manpages/man1/afptest.1.md:2
+#: manpages/man1/asip-status.1.md:2 manpages/man1/dbd.1.md:2
+#: manpages/man1/getzones.1.md:2 manpages/man1/macusers.1.md:2
+#: manpages/man1/nbp.1.md:2 manpages/man1/pap.1.md:2
+#: manpages/man1/rtmpqry.1.md:2 manpages/man3/atalk_aton.3.md:2
+#: manpages/man3/nbp_name.3.md:2 manpages/man4/atalk.4.md:2
+#: manpages/man5/afp_signature.conf.5.md:2
+#: manpages/man5/afp_voluuid.conf.5.md:2 manpages/man5/afp.conf.5.md:2
+#: manpages/man5/atalkd.conf.5.md:2 manpages/man5/extmap.conf.5.md:2
+#: manpages/man5/papd.conf.5.md:2 manpages/man8/a2boot.8.md:2
+#: manpages/man8/afpd.8.md:2 manpages/man8/atalkd.8.md:2
+#: manpages/man8/cnid_dbd.8.md:2 manpages/man8/cnid_metad.8.md:2
+#: manpages/man8/macipgw.8.md:2 manpages/man8/netatalk.8.md:2
+#: manpages/man8/papd.8.md:2 manpages/man8/papstatus.8.md:2
+#: manpages/man8/timelord.8.md:2
+#, fuzzy
+#| msgid "Name"
+msgid "# Name"
+msgstr "名前"
+
+#. type: Plain text
+#: manpages/man1/ad.1.md:6 manpages/man1/addump.1.md:6
+#: manpages/man1/aecho.1.md:6 manpages/man1/afp_lantest.1.md:6
+#: manpages/man1/afpldaptest.1.md:6 manpages/man1/afppasswd.1.md:6
+#: manpages/man1/afpstats.1.md:6 manpages/man1/afptest.1.md:6
+#: manpages/man1/asip-status.1.md:6 manpages/man1/dbd.1.md:6
+#: manpages/man1/getzones.1.md:6 manpages/man1/macusers.1.md:6
+#: manpages/man1/nbp.1.md:6 manpages/man1/pap.1.md:6
+#: manpages/man1/rtmpqry.1.md:6 manpages/man3/atalk_aton.3.md:6
+#: manpages/man3/nbp_name.3.md:6 manpages/man4/atalk.4.md:6
+#: manpages/man5/afp.conf.5.md:6 manpages/man8/a2boot.8.md:6
+#: manpages/man8/afpd.8.md:6 manpages/man8/atalkd.8.md:6
+#: manpages/man8/cnid_dbd.8.md:6 manpages/man8/cnid_metad.8.md:6
+#: manpages/man8/macipgw.8.md:6 manpages/man8/netatalk.8.md:6
+#: manpages/man8/papd.8.md:6 manpages/man8/papstatus.8.md:6
+#: manpages/man8/timelord.8.md:6
+#, fuzzy
+#| msgid "Synopsis"
+msgid "# Synopsis"
+msgstr "概要"
+
+#. type: Plain text
+#: manpages/man1/ad.1.md:12 manpages/man1/addump.1.md:20
+#: manpages/man1/aecho.1.md:10 manpages/man1/afp_lantest.1.md:11
+#: manpages/man1/afpldaptest.1.md:12 manpages/man1/afppasswd.1.md:10
+#: manpages/man1/afpstats.1.md:10 manpages/man1/afptest.1.md:23
+#: manpages/man1/asip-status.1.md:14 manpages/man1/dbd.1.md:12
+#: manpages/man1/getzones.1.md:10 manpages/man1/macusers.1.md:12
+#: manpages/man1/nbp.1.md:14 manpages/man1/pap.1.md:10
+#: manpages/man1/rtmpqry.1.md:10 manpages/man3/atalk_aton.3.md:15
+#: manpages/man3/nbp_name.3.md:17 manpages/man4/atalk.4.md:11
+#: manpages/man5/afp_signature.conf.5.md:6
+#: manpages/man5/afp_voluuid.conf.5.md:6 manpages/man5/atalkd.conf.5.md:6
+#: manpages/man5/extmap.conf.5.md:6 manpages/man5/papd.conf.5.md:6
+#: manpages/man8/a2boot.8.md:12 manpages/man8/afpd.8.md:12
+#: manpages/man8/atalkd.8.md:12 manpages/man8/cnid_dbd.8.md:12
+#: manpages/man8/cnid_metad.8.md:12 manpages/man8/macipgw.8.md:12
+#: manpages/man8/netatalk.8.md:12 manpages/man8/papd.8.md:12
+#: manpages/man8/papstatus.8.md:10 manpages/man8/timelord.8.md:12
+#, fuzzy
+#| msgid "Description"
+msgid "# Description"
+msgstr "説明"
+
+#. type: Plain text
+#: manpages/man1/addump.1.md:30 manpages/man1/aecho.1.md:41
+#: manpages/man1/afp_lantest.1.md:22 manpages/man1/afpldaptest.1.md:17
+#: manpages/man1/afppasswd.1.md:34 manpages/man1/asip-status.1.md:29
+#: manpages/man1/dbd.1.md:22 manpages/man1/getzones.1.md:17
+#: manpages/man1/macusers.1.md:16 manpages/man1/pap.1.md:30
+#: manpages/man1/rtmpqry.1.md:16 manpages/man8/a2boot.8.md:33
+#: manpages/man8/afpd.8.md:20 manpages/man8/atalkd.8.md:27
+#: manpages/man8/cnid_dbd.8.md:59 manpages/man8/cnid_metad.8.md:22
+#: manpages/man8/macipgw.8.md:39 manpages/man8/netatalk.8.md:27
+#: manpages/man8/papd.8.md:38 manpages/man8/papstatus.8.md:25
+#: manpages/man8/timelord.8.md:24
+#, fuzzy
+#| msgid "Options"
+msgid "# Options"
+msgstr "オプション"
+
+#. type: Plain text
+#: manpages/man1/addump.1.md:69 manpages/man1/aecho.1.md:47
+#: manpages/man1/afp_lantest.1.md:168 manpages/man1/afpldaptest.1.md:35
+#: manpages/man1/afppasswd.1.md:69 manpages/man1/afpstats.1.md:28
+#: manpages/man1/getzones.1.md:85 manpages/man1/macusers.1.md:26
+#: manpages/man1/pap.1.md:94 manpages/man1/rtmpqry.1.md:53
+#: manpages/man3/atalk_aton.3.md:25 manpages/man4/atalk.4.md:46
+#: manpages/man5/afp.conf.5.md:1349 manpages/man5/atalkd.conf.5.md:99
+#: manpages/man5/extmap.conf.5.md:29 manpages/man5/papd.conf.5.md:165
+#: manpages/man8/afpd.8.md:113 manpages/man8/atalkd.8.md:80
+#: manpages/man8/cnid_dbd.8.md:143 manpages/man8/cnid_metad.8.md:48
+#: manpages/man8/macipgw.8.md:99 manpages/man8/netatalk.8.md:58
+#: manpages/man8/papstatus.8.md:50
+#, fuzzy
+#| msgid "See Also"
+msgid "# See Also"
+msgstr "関連項目"
+
+#. type: Plain text
+#: manpages/man1/aecho.1.md:25 manpages/man1/afp_lantest.1.md:124
+#: manpages/man1/afppasswd.1.md:25 manpages/man1/afpstats.1.md:21
+#: manpages/man1/afptest.1.md:107 manpages/man1/asip-status.1.md:55
+#: manpages/man1/getzones.1.md:58 manpages/man1/nbp.1.md:63
+#: manpages/man1/rtmpqry.1.md:35 manpages/man3/nbp_name.3.md:27
+#: manpages/man5/afp_signature.conf.5.md:38
+#: manpages/man5/afp_voluuid.conf.5.md:28 manpages/man5/afp.conf.5.md:1303
+#: manpages/man5/atalkd.conf.5.md:81 manpages/man5/extmap.conf.5.md:19
+#: manpages/man5/papd.conf.5.md:92 manpages/man8/macipgw.8.md:79
+#, fuzzy
+#| msgid "Examples"
+msgid "# Examples"
+msgstr "例"
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:4
+msgid "afp_lantest — AFP LAN performance and directory cache testing tool"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:9
+#, no-wrap
+msgid ""
+"**afp_lantest** [-34567GgVvbC] [-h *host*] [-p *port*] [-s *volume*] [-u *user*] [-w *password*] \n"
+"[-n *iterations*] [-f *tests*] [-F *bigfile*]\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:15
+#, no-wrap
+msgid ""
+"**afp_lantest** is a comprehensive AFP (Apple Filing Protocol) performance testing tool designed to \n"
+"benchmark various aspects of AFP servers, with special focus on directory cache performance \n"
+"improvements in netatalk.\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:20
+msgid ""
+"The tool runs a series of tests that measure file operations, directory "
+"traversal, and caching efficiency. It includes both traditional file system "
+"benchmarks and specialized cache-focused tests that highlight the benefits "
+"of optimized directory cache validation and probabilistic validation "
+"features."
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:25
+#, no-wrap
+msgid ""
+"**-3**\n"
+": Use AFP 3.0 protocol version\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:28
+#, no-wrap
+msgid ""
+"**-4** \n"
+": Use AFP 3.1 protocol version\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:31
+#, no-wrap
+msgid ""
+"**-5**\n"
+": Use AFP 3.2 protocol version\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:34
+#, no-wrap
+msgid ""
+"**-6**\n"
+": Use AFP 3.3 protocol version\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:37
+#, no-wrap
+msgid ""
+"**-7**\n"
+": Use AFP 3.4 protocol version (default)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:40
+#, no-wrap
+msgid ""
+"**-h** *host*\n"
+": Server hostname or IP address (default: localhost)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:43
+#, no-wrap
+msgid ""
+"**-p** *port*\n"
+": Server port number (default: 548)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:46
+#, no-wrap
+msgid ""
+"**-s** *volume*\n"
+": Volume name to mount for testing\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:49
+#, no-wrap
+msgid ""
+"**-u** *user*\n"
+": Username for authentication (default: current uid)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:52
+#, fuzzy, no-wrap
+#| msgid "**-w** *password string*\n"
+msgid ""
+"**-w** *password*\n"
+": Password for authentication\n"
+msgstr "**-w** *パスワード文字列*\n"
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:55
+#, no-wrap
+msgid ""
+"**-n** *iterations*\n"
+": Number of test iterations to run (default: 1)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:58
+#, no-wrap
+msgid ""
+"**-f** *tests*\n"
+": Specific tests to run, specified as digits (e.g., \"134\" runs tests 1, 3, and 4)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:61
+#, no-wrap
+msgid ""
+"**-F** *bigfile*\n"
+": Use existing file in volume root for read test (file size must match -g/-G options)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:64
+#, no-wrap
+msgid ""
+"**-g**\n"
+": Optimize for gigabit networks (increases file test size to 1 GB)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:67
+#, no-wrap
+msgid ""
+"**-G**\n"
+": Optimize for 10-gigabit networks (increases file test size to 10 GB)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:70
+#, no-wrap
+msgid ""
+"**-C**\n"
+": Run cache-focused tests only (tests 8-11)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:73
+#, no-wrap
+msgid ""
+"**-v**\n"
+": Verbose output\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:76
+#, no-wrap
+msgid ""
+"**-V**\n"
+": Very verbose output\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:79
+#, no-wrap
+msgid ""
+"**-b**\n"
+": Debug mode\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:81
+#, fuzzy
+#| msgid "Available Commands"
+msgid "# Available Tests"
+msgstr "有効なコマンド"
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:83
+#, fuzzy
+#| msgid "The following FCE events are defined:"
+msgid "The following tests are available:"
+msgstr "以下の FCE イベントが定義されている:"
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:86
+#, no-wrap
+msgid ""
+"**(1) Opening, stating and reading 512 bytes from 1000 files**\n"
+": Tests basic file access patterns with many small operations\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:89
+#, no-wrap
+msgid ""
+"**(2) Writing one large file**\n"
+": Measures sustained write performance\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:92
+#, no-wrap
+msgid ""
+"**(3) Reading one large file**  \n"
+": Measures sustained read performance\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:95
+#, no-wrap
+msgid ""
+"**(4) Locking/Unlocking 10000 times each**\n"
+": Tests file locking performance\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:98
+#, no-wrap
+msgid ""
+"**(5) Creating dir with 2000 files**\n"
+": Tests directory creation and file creation performance\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:101
+#, no-wrap
+msgid ""
+"**(6) Enumerate dir with 2000 files**\n"
+": Tests directory enumeration with many files\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:104
+#, no-wrap
+msgid ""
+"**(7) Create directory tree with 10^3 dirs**\n"
+": Tests nested directory creation\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:107
+#, no-wrap
+msgid ""
+"**(8) Directory cache hits [CACHE]**\n"
+": Tests directory and file lookup performance (1000 dirs + 10000 files)\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:110
+#, no-wrap
+msgid ""
+"**(9) Mixed cache operations [CACHE]**\n"
+": Tests mixed create/stat/enum/delete operations\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:113
+#, no-wrap
+msgid ""
+"**(10) Deep path traversal [CACHE]**\n"
+": Tests performance with deep directory structures  \n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:116
+#, no-wrap
+msgid ""
+"**(11) Cache validation [CACHE]**\n"
+": Tests directory cache validation mechanisms\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:118
+msgid "## Cache-Focused Tests"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:122
+msgid ""
+"Tests 8-11 are specifically designed to highlight directory cache "
+"performance improvements in netatalk. These tests benefit significantly from "
+"optimized cache validation and probabilistic validation features. Use the **-"
+"C** option to run only these cache-focused tests."
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:129
+msgid ""
+"Run all tests with default settings: ``` afp_lantest -h server.example.com "
+"-u testuser -w password -s TestVolume ```"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:134
+msgid ""
+"Run only cache-focused tests: ``` afp_lantest -C -h server.example.com -u "
+"testuser -w password -s TestVolume ```"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:139
+msgid ""
+"Run specific tests (file operations): ``` afp_lantest -f 123 -h "
+"server.example.com -u testuser -w password -s TestVolume ```"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:144
+msgid ""
+"Run tests optimized for gigabit network: ``` afp_lantest -g -h "
+"server.example.com -u testuser -w password -s TestVolume ```"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:149
+msgid ""
+"Run multiple iterations for statistical analysis: ``` afp_lantest -n 5 -h "
+"server.example.com -u testuser -w password -s TestVolume ```"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:151
+#, fuzzy
+#| msgid "Return codes"
+msgid "# Return Codes"
+msgstr "復帰コード"
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:154
+#, no-wrap
+msgid ""
+"**0**\n"
+": All tests passed successfully\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:157
+#, no-wrap
+msgid ""
+"**1** \n"
+": One or more tests failed\n"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:159
+#, fuzzy
+#| msgid "Note"
+msgid "# Notes"
+msgstr "注記"
+
+#. type: Bullet: '- '
+#: manpages/man1/afp_lantest.1.md:166
+msgid ""
+"Tests that involve large file operations (tests 2 and 3) will create "
+"temporary files in the test volume"
+msgstr ""
+
+#. type: Bullet: '- '
+#: manpages/man1/afp_lantest.1.md:166
+msgid ""
+"The **-g** and **-G** options significantly increase test file sizes and "
+"test duration"
+msgstr ""
+
+#. type: Bullet: '- '
+#: manpages/man1/afp_lantest.1.md:166
+msgid ""
+"Cache-focused tests (8-11) provide the most insight into netatalk's "
+"directory cache performance"
+msgstr ""
+
+#. type: Bullet: '- '
+#: manpages/man1/afp_lantest.1.md:166
+msgid ""
+"Multiple iterations (**-n**) are recommended for consistent performance "
+"measurements"
+msgstr ""
+
+#. type: Bullet: '- '
+#: manpages/man1/afp_lantest.1.md:166
+msgid "The tool requires write access to the specified AFP volume"
+msgstr ""
+
+#. type: Plain text
+#: manpages/man1/afp_lantest.1.md:169
+#, no-wrap
+msgid "**afptest**(1), **afpd**(8), **netatalk**(8)\n"
+msgstr ""