--- a/doc/po4a.cfg
+++ b/doc/po4a.cfg
@@ -6,6 +6,7 @@
 [type: text] manpages/man1/ad.1.md $lang:translated/$lang/ad.1.md
 [type: text] manpages/man1/addump.1.md $lang:translated/$lang/addump.1.md
 [type: text] manpages/man1/aecho.1.md $lang:translated/$lang/aecho.1.md
+[type: text] manpages/man1/afp_lantest.1.md $lang:translated/$lang/afp_lantest.1.md
 [type: text] manpages/man1/afpldaptest.1.md $lang:translated/$lang/afpldaptest.1.md
 [type: text] manpages/man1/afppasswd.1.md $lang:translated/$lang/afppasswd.1.md
 [type: text] manpages/man1/afpstats.1.md $lang:translated/$lang/afpstats.1.md