… doc/manpages/man1/afptest.1.md
… test/testsuite/lantest.c