Add dircache-focused performance tests and enhance statistical analysis with robustness improvements to afp_lantest

    - Added 4 new cache performance tests (Tests 8-11): - TEST_CACHE_HITS: Directory cache efficiency testing - TEST_MIXED_CACHE_OPS: Cache lifecycle operations testing - TEST_DEEP_TRAVERSAL: Deep directory structure navigation - TEST_CACHE_VALIDATION: Cache validation efficiency testing
    - New -C command-line flag: Run cache-focused tests only
    - Enhanced analysis: Added standard deviation calculations with ± display format
    - Robustness improvements: Enhanced memory allocation, pthread handling, buffer safety
    - Bug fixes: Fixed array indexing, improved iteration tracking (off by one), enhanced error reporting (without debug mode)
    - Code quality: Added math.h header, improved results formatting, enhanced help text
    - Docs: Added new afp_lantest manpage

    Impact: afp_lantest enhancement adding new cache performance benchmarking capabilities
    while improving code reliability, statistical analysis, and user experience.

```
➜  /Volumes/case-sensitive/github/netatalk git:(Lantest) ✗ docker build -t netatalk-test:latest .
[+] Building 16.1s (17/17) FINISHED                                                                                                                                                                                                                                                                                                                  docker:desktop-linux
 => [internal] load build definition from Dockerfile                                                                                                                                                                                                                                                                                                                 0.0s
 => => transferring dockerfile: 3.25kB                                                                                                                                                                                                                                                                                                                               0.0s
 => [internal] load metadata for docker.io/library/alpine:3.22@sha256:4bcff63911fcb4448bd4fdacec207030997caf25e9bea4045fa6c8c44de311d1                                                                                                                                                                                                                               2.0s
 => [internal] load .dockerignore                                                                                                                                                                                                                                                                                                                                    0.0s
 => => transferring context: 80B                                                                                                                                                                                                                                                                                                                                     0.0s
 => [internal] load build context                                                                                                                                                                                                                                                                                                                                    0.3s
 => => transferring context: 19.43MB                                                                                                                                                                                                                                                                                                                                 0.3s
 => CACHED [build 1/6] FROM docker.io/library/alpine:3.22@sha256:4bcff63911fcb4448bd4fdacec207030997caf25e9bea4045fa6c8c44de311d1                                                                                                                                                                                                                                    0.0s
 => CACHED [build 2/6] RUN apk update &&  apk add --no-cache         avahi     cups     db     dbus     glib     iniparser     krb5     libevent     libgcrypt     linux-pam     mariadb-client     mariadb-connector-c     openldap     sqlite     tzdata              avahi-dev     cups-dev     db-dev     dbus-dev     build-base     gcc     iniparser-dev      0.0s
 => CACHED [build 3/6] WORKDIR /netatalk-code                                                                                                                                                                                                                                                                                                                        0.0s
 => [build 4/6] COPY . .                                                                                                                                                                                                                                                                                                                                             0.3s
 => [build 5/6] RUN meson setup build     -Dbuildtype=release     -Dwith-acls=false     -Dwith-afpstats=false     -Dwith-appletalk=true     -Dwith-testsuite=true     -Dwith-dbus-daemon-path=/usr/bin/dbus-daemon     -Dwith-dbus-sysconf-path=/etc     -Dwith-docs=     -Dwith-dtrace=false     -Dwith-init-style=none     -Dwith-pkgconfdir-path=/etc/netatalk    7.4s
 => [build 6/6] RUN meson install --destdir=/staging/ -C build                                                                                                                                                                                                                                                                                                       0.3s
 => [deploy 2/7] COPY --from=build /staging/ /                                                                                                                                                                                                                                                                                                                       0.0s
 => [deploy 3/7] COPY --from=build /netatalk-code/contrib/scripts/config_watch.sh /config_watch.sh                                                                                                                                                                                                                                                                   0.0s
 => [deploy 4/7] RUN apk update &&  apk add --no-cache     avahi     cups     db     dbus     glib     iniparser     krb5     libevent     libgcrypt     linux-pam     mariadb-client     mariadb-connector-c     openldap     sqlite     tzdata                                                                                                                     5.1s
 => [deploy 5/7] RUN ln -sf /dev/stdout /var/log/afpd.log                                                                                                                                                                                                                                                                                                            0.1s
 => [deploy 6/7] COPY /contrib/scripts/netatalk_container_entrypoint.sh /entrypoint.sh                                                                                                                                                                                                                                                                               0.0s
 => [deploy 7/7] WORKDIR /mnt                                                                                                                                                                                                                                                                                                                                        0.1s
 => exporting to image                                                                                                                                                                                                                                                                                                                                               0.2s
 => => exporting layers                                                                                                                                                                                                                                                                                                                                              0.2s
 => => writing image sha256:d873d6484e3fdbf6c513b6bc4df0ceb33b26623e4b4cd8e3cdbed2027a279a17                                                                                                                                                                                                                                                                         0.0s
 => => naming to docker.io/library/netatalk-test:latest                                                                                                                                                                                                                                                                                                              0.0s

➜  /Volumes/case-sensitive/github/netatalk git:(Lantest) ✗ docker run -d --network host --cap-add=NET_ADMIN --volume "/Volumes/case-sensitive/test:/mnt/afpshare" --volume "/Volumes/case-sensitive/test:/mnt/afpbackup" --env AFP_USER=test --env AFP_PASS=test --env SHARE_NAME='File Sharing' --env AFP_VERSION=7 --env AFP_GROUP=afpusers --env TZ=Europe/London --env INSECURE_AUTH=true --name netatalk-test netatalk-test:latest
e42f82bfb344aee9c683490bc3aeef0c38815b0cdeb82a17027fcc0be5fcf6dd

➜  /Volumes/case-sensitive/github/netatalk git:(Lantest) ✗ docker exec -it netatalk-test /usr/local/bin/afp_lantest -7 -h localhost -p 548 -u test -w test -s "File Sharing" -n 3
Run 0 => Opening, stating and reading 512 bytes from 1000 files   3588 ms
Run 0 => Writing one large file                                    173 ms for 100 MB (avg. 606 MB/s)
Run 0 => Reading one large file                                     45 ms for 100 MB (avg. 2330 MB/s)
Run 0 => Locking/Unlocking 10000 times each                        815 ms
Run 0 => Creating dir with 2000 files                             4499 ms
Run 0 => Enumerate dir with 2000 files                             808 ms
Run 0 => Create directory tree with 10^3 dirs                     1854 ms
Run 0 => Directory cache hits (1000 dir + 10000 file lookups)     3181 ms
Run 0 => Mixed cache operations (create/stat/enum/delete)          822 ms
Run 0 => Deep path traversal (nested directory navigation)         225 ms
Run 0 => Cache validation efficiency (metadata changes)           9394 ms
Run 1 => Opening, stating and reading 512 bytes from 1000 files   4057 ms
Run 1 => Writing one large file                                    122 ms for 100 MB (avg. 859 MB/s)
Run 1 => Reading one large file                                     52 ms for 100 MB (avg. 2016 MB/s)
Run 1 => Locking/Unlocking 10000 times each                        763 ms
Run 1 => Creating dir with 2000 files                             4193 ms
Run 1 => Enumerate dir with 2000 files                            1232 ms
Run 1 => Create directory tree with 10^3 dirs                     1877 ms
Run 1 => Directory cache hits (1000 dir + 10000 file lookups)     3197 ms
Run 1 => Mixed cache operations (create/stat/enum/delete)          690 ms
Run 1 => Deep path traversal (nested directory navigation)         253 ms
Run 1 => Cache validation efficiency (metadata changes)           9134 ms
Run 2 => Opening, stating and reading 512 bytes from 1000 files   3550 ms
Run 2 => Writing one large file                                    210 ms for 100 MB (avg. 499 MB/s)
Run 2 => Reading one large file                                     41 ms for 100 MB (avg. 2557 MB/s)
Run 2 => Locking/Unlocking 10000 times each                        800 ms
Run 2 => Creating dir with 2000 files                             3781 ms
Run 2 => Enumerate dir with 2000 files                             719 ms
Run 2 => Create directory tree with 10^3 dirs                     1837 ms
Run 2 => Directory cache hits (1000 dir + 10000 file lookups)     3759 ms
Run 2 => Mixed cache operations (create/stat/enum/delete)          495 ms
Run 2 => Deep path traversal (nested directory navigation)         426 ms
Run 2 => Cache validation efficiency (metadata changes)           9419 ms

Netatalk Lantest Results (average times across 3 iterations)
===================================

Test 1: Opening, stating and reading 512 bytes from 1000 files
	  Average:   3731 ms ± 282.4 ms (std dev)

Test 2: Writing one large file
	  Average:    168 ms ± 44.2 ms (std dev)
	  Throughput: 624 MB/s (Write, 100 MB file)

Test 3: Reading one large file
	  Average:     46 ms ± 5.6 ms (std dev)
	  Throughput: 2279 MB/s (Read, 100 MB file)

Test 4: Locking/Unlocking 10000 times each
	  Average:    792 ms ± 26.8 ms (std dev)

Test 5: Creating dir with 2000 files
	  Average:   4157 ms ± 360.3 ms (std dev)

Test 6: Enumerate dir with 2000 files
	  Average:    919 ms ± 274.1 ms (std dev)

Test 7: Create directory tree with 10^3 dirs
	  Average:   1856 ms ± 20.1 ms (std dev)

Test 8: Directory cache hits (1000 dir + 10000 file lookups)
	  Average:   3379 ms ± 329.2 ms (std dev)

Test 9: Mixed cache operations (create/stat/enum/delete)
	  Average:    669 ms ± 164.5 ms (std dev)

Test 10: Deep path traversal (nested directory navigation)
	  Average:    301 ms ± 108.9 ms (std dev)

Test 11: Cache validation efficiency (metadata changes)
	  Average:   9315 ms ± 157.8 ms (std dev)
```