@andylemin 2025-08-04
Done. I have removed all .po files other than the new one, and fixed a silent regression in testsuite_alp.Dockerfile which happened during my commit recovery. Looking better to me now